Moved files to query subpackage
