+++ genomics_data_index/api/query/impl/DataFrameSamplesQuery.py
@@ -5,8 +5,8 @@ from typing import Union, List
 import pandas as pd
 
 from genomics_data_index.api.SamplesQuery import SamplesQuery
-from genomics_data_index.api.impl.TreeSamplesQuery import TreeSamplesQuery
-from genomics_data_index.api.impl.WrappedSamplesQuery import WrappedSamplesQuery
+from genomics_data_index.api.query.impl.TreeSamplesQuery import TreeSamplesQuery
+from genomics_data_index.api.query.impl.WrappedSamplesQuery import WrappedSamplesQuery
 from genomics_data_index.configuration.connector import DataIndexConnection
 from genomics_data_index.storage.SampleSet import SampleSet
 from genomics_data_index.storage.model.QueryFeature import QueryFeature
