+++ genomics_data_index/api/query/impl/SamplesQueryIndex.py
@@ -5,9 +5,9 @@ from typing import Union, List, Set
 import pandas as pd
 
 from genomics_data_index.api.SamplesQuery import SamplesQuery
-from genomics_data_index.api.impl.DataFrameSamplesQuery import DataFrameSamplesQuery
-from genomics_data_index.api.impl.QueriesCollection import QueriesCollection
-from genomics_data_index.api.impl.TreeSamplesQuery import TreeSamplesQuery
+from genomics_data_index.api.query.impl.DataFrameSamplesQuery import DataFrameSamplesQuery
+from genomics_data_index.api.query.impl.QueriesCollection import QueriesCollection
+from genomics_data_index.api.query.impl.TreeSamplesQuery import TreeSamplesQuery
 from genomics_data_index.configuration.connector import DataIndexConnection
 from genomics_data_index.storage.SampleSet import SampleSet
 from genomics_data_index.storage.model.QueryFeature import QueryFeature
