+++ genomics_data_index/storage/service/SampleService.py
@@ -2,7 +2,7 @@ from typing import List, Dict, Set, Union, cast
 
 import pandas as pd
 
-from genomics_data_index.api.impl.QueriesCollection import QueriesCollection
+from genomics_data_index.api.query.impl.QueriesCollection import QueriesCollection
 from genomics_data_index.storage.SampleSet import SampleSet
 from genomics_data_index.storage.model.NucleotideMutationTranslater import NucleotideMutationTranslater
 from genomics_data_index.storage.model.QueryFeature import QueryFeature
